pronuncia-ia/requirements.txt
# ====================================
# Dependências do Projeto C317 - IA
# ====================================

# Framework Web
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Machine Learning & IA
transformers==4.44.2
torch==2.4.1
torchaudio==2.4.1
librosa==0.10.2
soundfile==0.12.1

# Modelos de Speech-to-Text
faster-whisper==1.0.3
# deepspeech==0.9.3  # Opcional, descomentar se usar
# coqui-stt==1.4.0   # Opcional, descomentar se usar

# APIs de IA
openai>=1.0.0
google-generativeai>=0.8.0

# Algoritmos de Similaridade
python-Levenshtein==0.25.1

# Utilitários
python-dotenv==1.0.1
python-multipart==0.0.9
numpy==1.26.4

# Testes
pytest==8.3.3
pytest-cov==5.0.0
httpx==0.27.2

# Opcional: para otimização
# accelerate==0.34.0  # Para GPU
# cuda-python==12.6.0  # Se tiver CUDA
